requirements.txt
# Account Discovery Tool - Python Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.25.0
playwright>=1.40.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
aiofiles>=23.2.1
